To get started with TensorFlow we ask you to specify a very simple computational graph of the function f(x)=x2+2x+5f(x)=x2+2x+5 and also taking its derivative. If you don't know how to proceed, please study the example given on the lecture and also the official TenorFlow starting guide: https://www.tensorflow.org/get_started/get_started.
This exercise requires around 10 lines of code, it is simple enough. However, it is very important to understand the notion of computational graph.

Actually, this simple example will be a backbone for all neural networks in the future. All what we will need is to specify first the forward pass (f), and then take derivatives with respect to parameters (x), in order to update them, so they minimize the loss function f. But we will come back to this in several weeks.